backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
gitpython==3.1.40
pytest==7.4.3
pytest-asyncio==0.21.1
psycopg2-binary==2.9.9
torch==2.1.1
numpy==1.24.3 